fix: define getAssetClient logic in only one function
