+++ src/lib/web3/hooks/useDenomClients.ts
@@ -4,51 +4,17 @@ import { useEffect, useMemo } from 'react';
 import { useDeepCompareMemoize } from 'use-deep-compare-effect';
 import {
   ChainRegistryClient,
-  ChainRegistryClientOptions,
   ChainRegistryChainUtil,
 } from '@chain-registry/client';
 
-import { Asset, AssetList, Chain, IBCInfo } from '@chain-registry/types';
+import { Asset, Chain } from '@chain-registry/types';
 import { DenomTrace } from '@duality-labs/dualityjs/types/codegen/ibc/applications/transfer/v1/transfer';
 
 import { useDenomTrace, useDenomTraceByDenom } from './useDenomsFromChain';
 import { Token } from '../utils/tokens';
+import { getAssetClient } from './useDenomsFromRegistry';
 
-const {
-  REACT_APP__CHAIN_NAME = '',
-  REACT_APP__CHAIN_IS_TESTNET = '',
-  REACT_APP__CHAIN_REGISTRY_FILE_ENDPOINTS = '["https://registry.ping.pub"]',
-  REACT_APP__CHAIN_REGISTRY_PATH_ENDPOINTS = '["https://registry.ping.pub"]',
-  REACT_APP__CHAIN_REGISTRY_ASSET_LISTS = '',
-  REACT_APP__CHAIN_REGISTRY_CHAINS = '',
-  REACT_APP__CHAIN_REGISTRY_IBC_DATA = '',
-} = import.meta.env;
-
-const isTestnet = REACT_APP__CHAIN_IS_TESTNET
-  ? REACT_APP__CHAIN_IS_TESTNET === 'true'
-  : REACT_APP__CHAIN_NAME.endsWith('testnet');
-
-// chain registry endpoints are endpoints where we can get ~up-to-date (cached)
-// asset and IBC data. it is expected that this data will change infrequently
-const chainRegistryFileEndpoints: string[] = JSON.parse(
-  REACT_APP__CHAIN_REGISTRY_FILE_ENDPOINTS
-).map((endpoint: string) => (isTestnet ? `${endpoint}/testnets` : endpoint));
-const chainRegistryDirectoryEndpoints: string[] = JSON.parse(
-  REACT_APP__CHAIN_REGISTRY_PATH_ENDPOINTS
-).map((endpoint: string) => (isTestnet ? `${endpoint}/testnets` : endpoint));
-
-// default chain-registry data can be used to supplement chain-registry data
-const defaultAssetLists: AssetList[] = REACT_APP__CHAIN_REGISTRY_ASSET_LISTS
-  ? JSON.parse(REACT_APP__CHAIN_REGISTRY_ASSET_LISTS)
-  : undefined;
-const defaultChains: Chain[] = REACT_APP__CHAIN_REGISTRY_CHAINS
-  ? JSON.parse(REACT_APP__CHAIN_REGISTRY_CHAINS)
-  : undefined;
-const defaultIbcData: IBCInfo[] = REACT_APP__CHAIN_REGISTRY_IBC_DATA
-  ? JSON.parse(REACT_APP__CHAIN_REGISTRY_IBC_DATA)
-  : undefined;
-
-type ChainNamePair = [chainName1: string, chainName2: string];
+const { REACT_APP__CHAIN_NAME = '' } = import.meta.env;
 
 // export hook for getting a basic chain-registry client for a denom
 // note: the client can do more than chainUtil which uses native chain context
@@ -59,11 +25,7 @@ export function useAssetClient(denom: string | undefined) {
     ['asset-client', denom, trace],
     async (): Promise<ChainRegistryClient | undefined> => {
       // get asset client if available
-      return (
-        (denom &&
-          getAssetClient(denom, trace || { path: '', base_denom: denom })) ||
-        undefined
-      );
+      return (denom && getAssetClient(denom, trace)) || undefined;
     }
   );
 }
@@ -111,15 +73,7 @@ export function useAssetClientByDenom(
       return denom ? [denom, trace, `asset-client-${traceKeys}`] : null;
     },
     async ([denom, trace]) => {
-      return [
-        denom,
-        denom
-          ? await getAssetClient(
-              denom,
-              trace || { path: '', base_denom: denom }
-            )
-          : undefined,
-      ];
+      return [denom, await getAssetClient(denom, trace)];
     },
     {
       parallel: true,
@@ -240,219 +194,4 @@ export function useToken(denom: string | undefined): SWRCommon<Token> {
   }, [tokenByDenom, denom]);
 
   return { ...swr, data };
-}
-
-// add a global fetch cache for chain-registry data
-// this will work like react-query and swr to prevent multiple fetch queries
-// from being sent out when the request is initialized at the same time
-// the cache should return any in-progress promises for duplicate URLs
-// note: the data here cannot be invalidated and re-requested, which is ok here
-const globalFetchPromises = new Map<string, Promise<unknown>>();
-function globalFetchCache(url: string): Promise<unknown> {
-  const oldPromise = globalFetchPromises.get(url);
-  if (oldPromise) {
-    return oldPromise;
-  }
-  const newPromise = fetch(url).then((res) => {
-    if (res.status >= 400) {
-      throw new Error('Bad response');
-    }
-    return res.json();
-  });
-  globalFetchPromises.set(url, newPromise);
-
-  return newPromise;
-}
-
-// recursively fetch enough chain-registry data to return a chain-registry
-// client that can identify the denom: eg. `chainUtil.getAssetByDenom(denom)`
-async function getAssetClient(
-  denom: string,
-  ibcTrace: DenomTrace
-): Promise<ChainRegistryClient | undefined> {
-  const transferChannels: Array<[portId: string, channelId: string]> =
-    ibcTrace.path
-      .split('/')
-      .flatMap((path, index, paths) =>
-        index % 2 !== 0 ? [[paths[index - 1], path]] : []
-      );
-
-  async function getClient(
-    opts: ChainRegistryClientOptions
-  ): Promise<ChainRegistryClient> {
-    for (const endpoint of chainRegistryFileEndpoints) {
-      try {
-        const client = new ChainRegistryClient({
-          ...opts,
-          baseUrl: endpoint,
-        });
-        // update client with cached or uncached fetch data
-        await Promise.all(
-          client.urls.map((url) =>
-            globalFetchCache(url).then((data) => {
-              // find matching data to merge
-              const assetListOrChain = data as AssetList | Chain;
-              if (assetListOrChain.chain_name === REACT_APP__CHAIN_NAME) {
-                // merge current chain asset data with default chain asset data
-                const assetList = data as AssetList;
-                if (
-                  defaultAssetLists &&
-                  assetList.$schema?.endsWith('/assetlist.schema.json')
-                ) {
-                  const defaultAssetList = defaultAssetLists.find(
-                    ({ chain_name }) => chain_name === assetList.chain_name
-                  )?.assets;
-                  return client.update({
-                    ...assetList,
-                    // add matched chain name assets over default assets by base
-                    assets: assetList.assets.reduce<Asset[]>((acc, asset) => {
-                      const found = acc.find(({ base }) => base === asset.base);
-                      // merge asset
-                      if (found) {
-                        acc[acc.indexOf(found)] = { ...found, ...asset };
-                      }
-                      // or add asset
-                      else {
-                        acc.push(asset);
-                      }
-                      return acc;
-                    }, defaultAssetList?.slice() ?? []),
-                  });
-                }
-                // merge current chain data with default chain data
-                const chain = data as Chain;
-                if (
-                  defaultChains &&
-                  chain.$schema?.endsWith('/chain.schema.json')
-                ) {
-                  const defaultChain = defaultChains.find(
-                    ({ chain_name }) => chain_name === chain.chain_name
-                  );
-                  return client.update({ ...defaultChain, ...chain });
-                }
-              }
-              return client.update(data as Chain | AssetList | IBCInfo);
-            })
-          )
-        );
-        return client;
-      } catch (e) {
-        // eslint-disable-next-line no-console
-        console.error(e);
-        // try next endpoint if there is one
-      }
-    }
-    throw new Error('Could not get current chain-registry data');
-  }
-
-  async function getIbcNamePairs(): Promise<ChainNamePair[]> {
-    for (const chainRegistryEndpoint of chainRegistryDirectoryEndpoints) {
-      try {
-        const data = await globalFetchCache(`${chainRegistryEndpoint}/_IBC`);
-        const json = data as Array<{ name: string }>;
-        return (
-          json.flatMap(({ name }: { name: string }) => {
-            const chainNames = name.split('.')?.at(0)?.split('-');
-            if (chainNames && chainNames[0] && chainNames[1]) {
-              return [[chainNames[0], chainNames[1]]];
-            }
-            return [];
-          }) || []
-        );
-      } catch (e) {
-        // eslint-disable-next-line no-console
-        console.error(e);
-        // try next endpoint if there is one
-      }
-    }
-    throw new Error('Could not get current chain-registry data');
-  }
-
-  // get IBC name pairs related to a chain
-  async function getRelatedIbcNamePairs(exploreChainName: string) {
-    const allIbcNamePairs = await getIbcNamePairs();
-    return exploreChainName
-      ? allIbcNamePairs?.filter((ibcPairChainNames) =>
-          ibcPairChainNames.includes(exploreChainName)
-        )
-      : undefined;
-  }
-
-  async function getAssetClient(
-    opts: ChainRegistryClientOptions
-  ): Promise<ChainRegistryClient | undefined> {
-    const client = await getClient(opts);
-    try {
-      // return successfully if we found the asset
-      const chainUtil = client.getChainUtil(REACT_APP__CHAIN_NAME);
-      if (chainUtil.getAssetByDenom(denom)) {
-        return client;
-      }
-    } catch {
-      // if the asset was not found, then determine if we should fetch more data
-      const lastChainName = opts.chainNames.at(-1) || '';
-      const ibcData = client.getChainIbcData(lastChainName);
-      const nextTransferChannelIndex = opts.chainNames.length - 1;
-      const nextTransferChannel = transferChannels.at(nextTransferChannelIndex);
-
-      // if we don't have all the channel hops covered then fetch more IBC data
-      if (ibcData && lastChainName && nextTransferChannel) {
-        const [portId, channelId] = nextTransferChannel;
-        for (const ibcDataRow of ibcData) {
-          // look up chain 1
-          if (ibcDataRow.chain_1.chain_name === lastChainName) {
-            const foundChannel = ibcDataRow.channels.find((channel) => {
-              return (
-                channel.chain_1.channel_id === channelId &&
-                channel.chain_1.port_id === portId
-              );
-            });
-            if (foundChannel) {
-              return getNextChainNameClient(ibcDataRow.chain_2.chain_name);
-            }
-          }
-          // look up chain 2
-          else if (ibcDataRow.chain_2.chain_name === lastChainName) {
-            const foundChannel = ibcDataRow.channels.find((channel) => {
-              return (
-                channel.chain_2.channel_id === channelId &&
-                channel.chain_2.port_id === portId
-              );
-            });
-            if (foundChannel) {
-              return getNextChainNameClient(ibcDataRow.chain_1.chain_name);
-            }
-          }
-        }
-      }
-    }
-
-    async function getNextChainNameClient(chainName: string) {
-      // if this would be the last step then find the assetlist of this chain
-      // otherwise we should keep searching IBC data
-
-      if (opts.chainNames.length >= transferChannels.length) {
-        return getAssetClient({
-          ...opts,
-          assetListNames: [REACT_APP__CHAIN_NAME, chainName],
-        });
-      } else {
-        return getAssetClient({
-          ...opts,
-          chainNames: [...opts.chainNames, chainName],
-          ibcNamePairs: await getRelatedIbcNamePairs(chainName),
-        });
-      }
-    }
-  }
-
-  // start recursive chain with just native assets and all related IBC info
-  return getAssetClient({
-    chainNames: [REACT_APP__CHAIN_NAME],
-    ibcNamePairs: await getRelatedIbcNamePairs(REACT_APP__CHAIN_NAME),
-    assetListNames: [REACT_APP__CHAIN_NAME],
-    assetLists: defaultAssetLists,
-    chains: defaultChains,
-    ibcData: defaultIbcData,
-  });
 }
